The "fourier_transform" and "fourier_transform_in_parallel" functions have been updated. The beginning of the implementation of "inverse_fourier_transform".
In the "fourier_transform" and "fourier_transform_in_parallel" functions, data return was implemented and the "amplitude" variable was added.

For the function "inverse_fourier_transform", the preparatory function "mirror" was implemented. The "mirror" function adds a mirror image of the Fourier transform array that was obtained using the "fourier_transform_in_parallel" or "fourier_transform" function.