Updated validation of input data for functions. Input data validations have been added for the functions. And one more thing.
Updated validation of input data path name or file name for functions:
    - "signal_recording";
    - "signal_playback";
    - "creating_a_signal".

Input data validations have been added for the functions:
    - "fourier_transform_in_parallel";
    - "fourier_transform";
    - "building_a_wave".

Updated the algorithm for preparing the filename and saving path for the graphs in the "building_a_wave" function.

A small fix in the function "building_a_fourier_transform_graph".